ENH : avoid segfault by casting the pointer in unsigned long cause in Win32 bad interpretation of the adress

+++ Code/IO/otbGDALImageIO.cxx
@@ -929,7 +929,7 @@ void GDALImageIO::Write(const void* buffer)
     // If JPEG, set the image quality
     if( gdalDriverShortName.compare("JPEG") == 0 )
       {
-  option[0] = const_cast<char *>("QUALITY=95");
+      option[0] = const_cast<char *>("QUALITY=95");
  
       }
     
@@ -1058,7 +1058,7 @@ void GDALImageIO::InternalWriteImageInformation(const void* buffer)
     {
     std::ostringstream stream;
     stream << "MEM:::"
-           <<  "DATAPOINTER=" << buffer << ","
+           <<  "DATAPOINTER=" << (unsigned long)(buffer) << ","
            <<  "PIXELS=" << m_Dimensions[0] << ","
            <<  "LINES=" << m_Dimensions[1] << ","
            <<  "BANDS=" << m_NbBands << ","
